GPS: Disbales code that completely shuts down our connection with GPS
This avoids a race condition that can crash the BP if the GPS is disabled
in settings while it is running. This also fixes a problem with the GPS icon
in the status bar not getting removed when the GPS is disabled.

(cherry picked from commit 5f2deb776d37d6ae2e02877f2160fc93dd33c8ae)

Change-Id: Ib20c448b20eab1bc32844d291c0dba4cc2389aec