Fixes BIT messaging and race conditions at power up
BIT used to request for IF bringup and then turn around connect,
not waiting for acknowlodgement from connectivity manager. We
added messaging between loc_eng and the daemon for so that the
daemon can wait for acknowledgement.
Also there is a race condition between the daemon and AFW at power
up, where when the daemon requests for IF, AFW hasn't init'ed
AGPS yet.

Change-Id: Iecbf16c875b5e869746cf785b4dedbbdc2fdc850
(merge of 1e847862cd0f35a3e9a7e12d663ad69bd6156a32)