App credentials support
Receive app credentials as part of the position mode parameters; and pass it
down to modem. There is also a slight change in the set_position_mode call
so that it receives a struct instead of individual params, as it is getting
too many of them. The 3rd change is removed the last position_mode caching
from loc_eng, and made LocApiAdapter do it, since the v02 adapter was already
doing it. So now the parent adapter does it for both v01 and v02 adapters.

Change-Id: I0426e67cdc59fdc603735fe1aacc1a2453a00f68