stopping a tracking session state flag may set wrong
When stopping a tracking session, stop_fix_handler sets navigating flag
to false when it exits. The engine off handling would also set the flag
but would set it to true if it is tracking session. This causes the next
tracking not starting, because it thinks it is already in session.

CRs-Fixed: 338534
(cherry picked from commit 6e4bfd682d336dce231fe4f1dffcea6eb3b00238)

Change-Id: Ia17be23abf48b0048529911bda02a65b1d9caa5d