stopping a tracking session state flag may set wrong
When stopping a tracking session, stop_fix_handler sets navigating flag
to false when it exits. The engine off handling would also set the flag
but would set it to true if it is tracking session. This causes the next
tracking not starting, because it thinks it is already in session.

Change-Id: Ie95875da3df91987993a63696d61a3682e24d996
CRs-Fixed: 338534