SSR bug
loc_clear did not reset userdata. loc_open then won't set the callbacks
in a strange combination of logics. It is also a potential hazard for
VZW devices in loc_open. With this change we will only check if userdata
is in the glue table to decide if some user has registered.

Change-Id: Ie540cacef0cd6b35b2d6c8840fccee87d2098a2a
CRs-Fixed: 343382