ATL call crash fix
AgpsStatus added with a couple string fields for wifi
which are not NULL ended when they are not used. This
get passed to AFW JNI, and allow UTF8 string with the
open ended buffer with junks in it. It might cause
exception depending on the content of the garbadge.

CRs-Fixed: 405473

(cherry picked from commit 67754118cec93d8032fda2e66901e157373ae407)

Change-Id: I200778ff840a08c47f26a9cde11a2720f23096de