Utility to identify target HW
Earlier this identification was done in loc.cpp
This change moves that to a separate utility.
Also included is a change to not return
a handle to the GPS interface if the target
is found to be MPQ8064.

CRs-Fixed: 418009
(cherry picked from commit 63c5ac48c7d0b9245abd1a5e3ab45c3f331655d5)

Change-Id: I831aabf261c0257494578695124c1ff0e95a1a82